backend/requirements.txt

fastapi==0.115.8
uvicorn[standard]==0.30.6
pydantic==2.8.2
PyYAML==6.0.2
